enhancement: made nav dynamic to stop it repeating links of the same page
