+++ public/chunk-map.json
@@ -1,1 +1,1 @@
-{"polyfill":["/polyfill.js","/polyfill.91bafec0c00b1be022d8.hot-update.js"],"commons":["/commons.css","/commons.js","/commons.91bafec0c00b1be022d8.hot-update.js"],"component---cache-dev-404-page-js":["/component---cache-dev-404-page-js.js"],"component---src-pages-404-js":["/component---src-pages-404-js.js","/component---src-pages-404-js.91bafec0c00b1be022d8.hot-update.js"],"component---src-pages-about-js":["/component---src-pages-about-js.js","/component---src-pages-about-js.91bafec0c00b1be022d8.hot-update.js"],"component---src-pages-contact-js":["/component---src-pages-contact-js.js","/component---src-pages-contact-js.91bafec0c00b1be022d8.hot-update.js"],"component---src-pages-index-js":["/component---src-pages-index-js.js","/component---src-pages-index-js.91bafec0c00b1be022d8.hot-update.js"],"component---src-pages-work-js":["/component---src-pages-work-js.js","/component---src-pages-work-js.91bafec0c00b1be022d8.hot-update.js"],"component---src-templates-work-project-work-js":["/component---src-templates-work-project-work-js.js","/component---src-templates-work-project-work-js.91bafec0c00b1be022d8.hot-update.js"]}
+{"polyfill":["/polyfill.js","/polyfill.6ad7aad8770ea8150853.hot-update.js"],"commons":["/commons.css","/commons.js","/commons.6ad7aad8770ea8150853.hot-update.js"],"component---cache-dev-404-page-js":["/component---cache-dev-404-page-js.js"],"component---src-pages-404-js":["/component---src-pages-404-js.js","/component---src-pages-404-js.6ad7aad8770ea8150853.hot-update.js"],"component---src-pages-about-js":["/component---src-pages-about-js.js","/component---src-pages-about-js.6ad7aad8770ea8150853.hot-update.js"],"component---src-pages-contact-js":["/component---src-pages-contact-js.js","/component---src-pages-contact-js.6ad7aad8770ea8150853.hot-update.js"],"component---src-pages-index-js":["/component---src-pages-index-js.js","/component---src-pages-index-js.6ad7aad8770ea8150853.hot-update.js"],"component---src-pages-work-js":["/component---src-pages-work-js.js","/component---src-pages-work-js.6ad7aad8770ea8150853.hot-update.js"],"component---src-templates-work-project-work-js":["/component---src-templates-work-project-work-js.js","/component---src-templates-work-project-work-js.6ad7aad8770ea8150853.hot-update.js"]}
